--- src/components/Search/SearchTypeSelector/SearchTypeSelector.jsx
@@ -1,37 +1,0 @@
-import React, { useState } from 'react';
-import { Wrapper } from './SearchTypeSelector.styled';
-import { Dropdown } from './Dropdown';
-
-export const SearchTypeSelector = ({ setTypeQuery }) => {
-  // eslint-disable-next-line no-unused-vars
-  const [searchType, setSearchType] = useState('query');
-
-  const handleSelectChange = selectedOption => {
-    setSearchType(selectedOption.type);
-    setTypeQuery(selectedOption.type);
-  };
-
-  const searchTypes = [
-    {
-      type: 'query',
-      label: 'Title',
-    },
-    {
-      type: 'ingredient',
-      label: 'Ingredients',
-    },
-  ];
-
-  return (
-    <>
-      <Wrapper>
-        <p>Search by:</p>
-        <Dropdown
-          onChange={handleSelectChange}
-          id={'selectSearchType'}
-          data={searchTypes}
-        />
-      </Wrapper>
-    </>
-  );
-};
