+++ src/index.js
@@ -30,11 +30,11 @@ const AppWrapper = () => {
 };
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <BrowserRouter basename="/recipes5">
-        <AppWrapper />
-      </BrowserRouter>
-    </Provider>
-  </React.StrictMode>
+  //<React.StrictMode>
+  <Provider store={store}>
+    <BrowserRouter basename="/recipes5">
+      <AppWrapper />
+    </BrowserRouter>
+  </Provider>
+  //</React.StrictMode>
 );
